+++ src/main/java/dev/willyelton/variable_redstone/common/block/VariableButton.java
@@ -1,0 +1,71 @@
+package dev.willyelton.variable_redstone.common.block;
+
+import dev.willyelton.variable_redstone.client.VariableScreen;
+import net.minecraft.client.Minecraft;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
+import net.minecraft.world.InteractionResult;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.level.BlockGetter;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.ButtonBlock;
+import net.minecraft.world.level.block.state.BlockBehaviour;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.state.StateDefinition;
+import net.minecraft.world.level.block.state.properties.AttachFace;
+import net.minecraft.world.level.block.state.properties.BlockSetType;
+import net.minecraft.world.level.gameevent.GameEvent;
+import net.minecraft.world.phys.BlockHitResult;
+import org.jetbrains.annotations.Nullable;
+
+import static dev.willyelton.variable_redstone.VariableRedstone.POWER;
+
+public class VariableButton extends ButtonBlock {
+    public VariableButton() {
+        super(BlockSetType.STONE, 0, BlockBehaviour.Properties.ofFullCopy(Blocks.STONE_BUTTON));
+        this.registerDefaultState(this.stateDefinition.any()
+                .setValue(FACING, Direction.NORTH)
+                .setValue(POWERED, false)
+                .setValue(FACE, AttachFace.WALL)
+                .setValue(POWER, 15));
+    }
+
+    @Override
+    protected InteractionResult useWithoutItem(BlockState state, Level level, BlockPos pos, Player player, BlockHitResult hitResult) {
+        if (player.isShiftKeyDown()) {
+            if (level.isClientSide) {
+                Minecraft.getInstance().setScreen(new VariableScreen(state.getValue(POWER), pos));
+            }
+            return InteractionResult.SUCCESS;
+        } else {
+            return super.useWithoutItem(state, level, pos, player, hitResult);
+        }
+    }
+
+    @Override
+    public void press(BlockState state, Level level, BlockPos pos, @Nullable Player player) {
+        level.setBlock(pos, state.setValue(POWERED, true), 3);
+        level.updateNeighborsAt(pos, this);
+        level.updateNeighborsAt(pos.relative(getConnectedDirection(state).getOpposite()), this);
+        level.scheduleTick(pos, this, 10);
+        this.playSound(player, level, pos, true);
+        level.gameEvent(player, GameEvent.BLOCK_ACTIVATE, pos);
+    }
+
+    @Override
+    protected int getSignal(BlockState blockState, BlockGetter blockAccess, BlockPos pos, Direction side) {
+        return blockState.getValue(POWERED) ? blockState.getValue(POWER) : 0;
+    }
+
+    @Override
+    protected int getDirectSignal(BlockState blockState, BlockGetter blockAccess, BlockPos pos, Direction side) {
+        return blockState.getValue(POWERED) && getConnectedDirection(blockState) == side ? blockState.getValue(POWER) : 0;
+    }
+
+    @Override
+    protected void createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> builder) {
+        builder.add(FACING, POWERED, FACE, POWER);
+    }
+}