+++ src/main/java/dev/willyelton/variable_redstone/datagen/VariableRedstoneItemTags.java
@@ -1,0 +1,4 @@
+package dev.willyelton.variable_redstone.datagen;
+
+public class VariableRedstoneItemTags {
+}