+++ src/main/java/dev/willyelton/variable_redstone/datagen/VariableRedstoneLootTables.java
@@ -19,6 +19,7 @@ public class VariableRedstoneLootTables extends VanillaBlockLoot {
         dropSelf(VariableRedstone.VARIABLE_REDSTONE_BLOCK.get());
         dropSelf(VariableRedstone.VARIABLE_REDSTONE_TORCH.get());
         dropSelf(VariableRedstone.VARIABLE_LEVER.get());
+        dropSelf(VariableRedstone.VARIABLE_BUTTON.get());
     }
 
     @Override